Updated All current json
Alls json in : 
root/OtterFloat/PrismaticJellyfish/Default/*
has been updated. 
Some more distinction might be needed for DialoguePanel and FloatPanel, as I found that they have the same state and function set. DialoguePanel was named as it is available to the User as a component to their creation as part of PrismaFish. FloatPanel was named because the DialoguePanel was decided to be the core feature to finish first in OtterFloat Edition. That is to say: FloatPanel is a Specific Type of DialoguePanel that is used primarily in the OtterFloat Editor. DialoguePanel would be the base class, and is a sophisticated base class for the User or Creator to customize for various purposes. A Panel can be thought of a like a Windowed-type Gui, a DialoguePanel is a Panel with Preset Data and Functions to imply its usages. Finally, a FloatPanel is a Dialogue Panel with more specific usage consequently has more specific Preset Data and Functions and maybe be used as discreet components to a more complicated OR more streamlined App/Game/Function/Tool.